refactor:코드 컨밴션에 맞게 수정
