+++ frontend/src/components/newTrip/NewTripForm/TripCreateForm.style.ts
