+++ frontend/src/hooks/api/useAddTripItemMutation.ts
@@ -10,7 +10,7 @@ export const useAddTripItemMutation = () => {
       // 순서 변경 성공 시 Trip 정보 재요청
       queryClient.invalidateQueries({ queryKey: ['trip'] });
     },
-    onError: (err, _, context) => {
+    onError: () => {
       alert('오류가 발생했습니다. 잠시 후 다시 시도해주세요.');
     },
   });
