+++ frontend/src/hooks/api/useTripEditMutation.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { putTrip } from '@api/trip/putTrip';
 
-export const useEditTripMutation = () => {
+export const useTripEditMutation = () => {
   const queryClient = useQueryClient();
 
   const tripMutation = useMutation(putTrip(), {
@@ -10,7 +10,8 @@ export const useEditTripMutation = () => {
       // 여행 정보 수정 성공시 Trip 정보 재요청
       queryClient.invalidateQueries({ queryKey: ['trip'] });
     },
-    onError: (err, _, context) => {
+    onError: () => {
+      //TODO: toast 띄우기
       alert('오류가 발생했습니다. 잠시 후 다시 시도해주세요.');
     },
   });
