+++ frontend/src/pages/TripCreatePage/TripCreatePage.tsx
@@ -1,3 +1,4 @@
+import TripCreateForm from '@/components/newTrip/NewTripForm/TripCreateForm';
 import CreatePageImage from '@assets/svg/create-page-image.svg';
 import { Box, Flex, Heading } from 'hang-log-design-system';
 
@@ -9,8 +10,6 @@ import {
   containerStyling,
 } from '@pages/TripCreatePage/TripCreatePage.style';
 
-import TripAddForm from '@components/newTrip/NewTripForm/NewTripForm';
-
 const TripCreatePage = () => {
   useCityQuery();
 
@@ -18,7 +17,7 @@ const TripCreatePage = () => {
     <Flex css={containerStyling}>
       <Box css={boxStyling}>
         <Heading>여행을 기록해 보세요</Heading>
-        <TripAddForm />
+        <TripCreateForm />
       </Box>
       <CreatePageImage aria-label="세계 랜드마크 이미지" css={backgroundImage} />
     </Flex>
